requirements.txt
fastapi
uvicorn
python-multipart
pydantic
requests
numpy
faiss-cpu
pypdf
python-dotenv
